Add favourite test + small refactor
Using 'favourite' rather than 'isFavourite' as a boolean name because
we're doing automatic mapping of API response data into the LinkModel
structure, so names need to match across the pair.

At least, that holds unless we want to just build a translation map
of names used by the API to names we use internally. Which may be
necessary in future, especially if more terminology has issues like
this. But for now I'm ok with it as is.